+++ internal/plugin/zz_provider.go
@@ -43,6 +43,7 @@ import (
 	"github.com/aiven/terraform-provider-aiven/internal/plugin/service/planlist"
 	"github.com/aiven/terraform-provider-aiven/internal/plugin/service/servicelist"
 	"github.com/aiven/terraform-provider-aiven/internal/plugin/service/staticip"
+	"github.com/aiven/terraform-provider-aiven/internal/plugin/service/upgradepipeline/step"
 )
 
 func Resources() map[string]func() resource.Resource {
@@ -73,6 +74,7 @@ func Resources() map[string]func() resource.Resource {
 		"aiven_pg_database":                         adapter.NewLazyResource(database2.ResourceOptions),
 		"aiven_pg_user":                             adapter.NewLazyResource(user3.ResourceOptions),
 		"aiven_static_ip":                           adapter.NewLazyResource(staticip.ResourceOptions),
+		"aiven_upgrade_step":                        adapter.NewLazyResource(step.ResourceOptions),
 	}
 }