+++ CHANGELOG.md
@@ -12,6 +12,7 @@ nav_order: 1
 
 ## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
 
+- Add `aiven_upgrade_step` resource: manages Upgrade Pipeline steps between source and destination services.
 - Add `aiven_clickhouse` field `clickhouse_user_config.private_access.clickhouse_arrowflight`: Allow clients to connect
   to clickhouse_arrowflight with a DNS name that always resolves to the service's private IP addresses
 - Add `aiven_clickhouse` field `clickhouse_user_config.privatelink_access.clickhouse_arrowflight`: Enable clickhouse_arrowflight.